requirements.txt
python-telegram-bot==20.6
flask==3.0.0
flask-cors==4.0.0
psycopg3[binary]==3.1.8
paypalrestsdk==1.13.3
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
gunicorn==21.2.0
